Extracted code from ojph_codestream_local.cpp
